+++ packages/terra-clinical-item-display/src/terra-dev-site/test/clinical-item-display/display/MeaningItemDisplay.test.jsx
@@ -4,9 +4,9 @@ import ItemDisplay from '../../../../ItemDisplay';
 
 const views = () => (
   <div>
-    <ItemDisplay text="test text" textStyle="attention" id="test-display-attention-meaning" meaning="Important" />
-    <ItemDisplay text="test text (default meaning)" textStyle="strikeThrough" id="test-display-strike-through-default" />
-    <ItemDisplay text="test text" textStyle="strikeThrough" id="test-display-strike-through-meaning" meaning="Update" />
+    <ItemDisplay text="test text" textStyle="attention" id="test-display-attention-meaning" textStyleMeaning="Important" />
+    <ItemDisplay text="test text (default)" textStyle="strikeThrough" id="test-display-strike-through-default" />
+    <ItemDisplay text="test text" textStyle="strikeThrough" id="test-display-strike-through-meaning" textStyleMeaning="Update" />
   </div>
 );