+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added `meaning` prop to allow defining a meaning for styled text to be read by screen readers.
+  * Added `textStyleMeaning` prop to allow defining a meaning for styled text to be read by screen readers.
 
 ## 4.6.0 - (March 29, 2023)